Reset supportsEnterPipOnTaskSwitch when paused
The ActivityRecord#supportsEnterPipOnTaskSwitch flag is to indicate the
Activity can enter PiP while pausing.

This flag is set to true in Transitions#checkEnterPipOnFinish and
wouldn't be reset back in certain scenarios. This can lead to issues
like accidentally enter PiP when do quick switch.

In this change, we make sure the flag is always set to false whenever
the activity is paused. We attempted to do so in makeInvisible if the
internal state is STOPPING/STOPPED, which can be skipped if the
mVisibleRequested flag is set to false first.

Flag: EXEMPT bugfix
Bug: 390872198
Video: http://recall/-/aaaaaabFQoRHlzixHdtY/gvYPNiIVWycrSYPQNZ0J9z
Test: manually make sure legacy enter PiP works
Test: follow the reproduce steps in bug, see also the video
Change-Id: I04ef9435e70f4d1ca41c7c5d7859d3c9462aa2aa

Former-commit-id: 2f425ca98a68826d2073ea4acd58deee0af87094