Apply task-limit policy in #moveToDisplay
Moving a task to a another displays means that if the display already
had desktop tasks at the limit, one of them must be minimized to make
room for the opening one.

The task-limit policy is already handled in addDeskActivationChanges,
but it relies on the |newTask| paremeter to indicate that a task is
moving into the desk along with the activation.

Flag: com.android.window.flags.enable_move_to_next_display_shortcut
Bug: 402144410
Test: Move task to other display with already open tasks at the limit,
verify the last task closes
Test: atest WMShellUnitTests

Change-Id: I4a26098ec2e386faf543011c655616df86518ee2

Former-commit-id: 9c4db782262df23192e92498b5be1b4bb1fcbfa2