Pass never-blank state from VirtualDisplayAdapter to SurfaceFlinger.
The never-blank concept is translated in VirtualDisplayAdapter to be a power-optimised display (instead of a performance optimised display) in the downstream components. This concept indicates that these displays (created via DisplayManager instead of through a virtual device) depend on another display's presence and activeness to be continually rendered and shown. The display they depend on instead optimises for performance when it is on, meaning that all the related displays have good performance.

For cases where dependent display/power optimisation information is not available (calls to DisplayControl#createVirtualDisplay(String, boolean) and SurfaceComposerClient::createVirtualDisplay), the dependent display/power optimisation state is defaulted to true since these are system calls and this preserves the previous behaviour.

Bug: 342681202

Flag: EXEMPT refactor

Test: Flashed build on test device, ran Android Auto Projected
Change-Id: I21d916635c161c3309928edc541e4ee0f3921b25

Former-commit-id: b989023ee03189eeca140dfb22180981ccb82897