Animate shade blur fallback
Instead of dispatch and apply state in ScrimController, we call startAnimation on each scrim with the current scrim tints saved before repreparing the states (to update colors based on blur state).

The transition works well in the unlocked state but more work possibly needs to be done for the bouncer.

Bug: 388891313
Test: manual
Flag: com.android.systemui.notification_shade_blur
Change-Id: I95d7dd0b5207f392620cf1d624b193bf3594e84b

Former-commit-id: e0677abea48704b418eec9aa7ea7d2c5f09518e3