Don't start binding to remote ExternalStorageService in some flows
In case we try to end fuse session or notifying volume state changes, we
try to bind even if we haven't managed to bind to the remote
ExternalStorageService before. In such cases, there is no need to bring
up the remote service as this situation means that we didn't actually
start a session before, or the service has died. In both of these cases,
when the remote service comes up later, we will make sure that
everything is reset, so that we start over the mounting process and
hence notify the correct state. The current logic will unnecessarily
wait for long time to bring the service up to notify stale events or end
a session that won't be there anyway as the remote service wasn't started
before or died at some time later.

Additionally, reduce the timeout time to 10 seconds in the above cases
as we no longer wait for service binding.

Bug: 369519866
Test: Manual & atest StorageManagerTest
Flag: EXEMPT bufix
Change-Id: I5233167e842ee8e8625d0c44e461b45a4577559c

Former-commit-id: 6068c5a5bfd76505bc588a46e02d861097bd77ac