Fix session start race condition
Session finished callback is triggered before the session is ended at
the VibratorManagerService, which can cause a race condition if a second
session is started right away.

This will cause the second session to be ignored because the service is
still cleaning up the previous session.

Fixing the callback to be triggered after all operations at the service
are complete and the session is trully finished.

Bug: 392420704
Test: android.os.cts.VendorVibrationSessionTest
Flag: android.os.vibrator.vendor_vibration_effects
Change-Id: Ia2f1c2dec29eb5bc7515e9de1bd236b77bd84b83

Former-commit-id: c01811dfddd693c0288ad3da1bee4307d96a2409