[battery] Don't create a BatteryStatusChip in tests
This battery chip view was used in such a way that over-constrained the
test itself (it was dependent on the implementation of BatteryStatusChip).

The new battery chip view is based on compose, and this test
realistically doesn't care about the underlying chip view, it's more
about testing the scheduler.

Therefore, when the flag is on we just create a generic view, and modify
a couple tests to reflect this.

Test: SystemStatusAnimationSchedulerImplTest
Bug: 391606042
Flag: com.android.settingslib.flags.new_status_bar_icons
Change-Id: Ife225794f50aa4f2f1de8ca4f22e8427c1cc0a71

Former-commit-id: 47b81434d737a684fc72fc750d8c45edad1d04f4