[PiP2] Handle legacy case if organized TF present
If an app uses AE, but its TF is effectively filling
the task or if it isn't really changing, it still might get
added into the changes list (task fragments can be organized which
are always added to the list of targets even if canPromote() returns
true).

So on Shell side we can effectively count such container "chains"
as one underlying change under a root PiP task. This way we can still
animate it as a normal single-activity legacy-enter PiP transition.

Bug: 398862854
Flag: com.android.wm.shell.enable_pip2
Test: swipe-to-enter PiP from a Whatsapp call
Change-Id: I8f06fdcb607680230ab31909113184f97de28252

Former-commit-id: 911848372d22513026644c28668b2f809e780102