[32/N] Desks: Use RunOnTransitStart to apply "pending" minimize/activations
This change moves both minimization (task-limit) and desk activation WCT
changes into a single, reusable util (#addDeskActivationChanges) and
updates callsites of both to use it. To accomodate both features having
to "wait" for the transition token to be available to add their "pending
state" in their respective observers, it changes the return value of the
util to a |RunOnTransitStart| callback that is invoked when the token is
ready.

This removes the burden on callers of knowing the specific side effects
of #addDeskActivationChanges and having to update the feature-specific
observers with the pending transition state.

Ideally, all other side-effects (e.g. immersive ones) should be migrated
into this method, but this CL focuses on just these two as a starting
point to unblock multi-desk development.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 362720497
Test: atest WMShellUnitTests
Change-Id: I9d95813271f65ea6f58af2cf20ff2ddcea54b52c

Former-commit-id: 0d3ebe8059be9bb9bbc04ec87f70d4cd3dea31b6