Fix bug where IME caused an unnecessary divider reposition in flex split
In ag/30137792, we introduced a way for the divider bar to animate to a new position when insets are changed, so that flex split ratios could adjust when changing to Pinned Taskbar or 3-button-nav.

However, this new function (`calculatePinnedTaskbarInsets`) mistakenly reported insets as zero when IME was active. Actually, IME insets are independent of pinned taskbar insets, so there was no need to make a special case for IME.

This patch fixes that by removing the IME check, allowing divider position to remain stable when IME appears and disappears.

Bug: 396031160
Bug: 399304764
Flag: com.android.wm.shell.enable_flexible_two_app_split
Test: IME launches properly and does not cause a divider reposition anymore.
Change-Id: Id74a38d75971213c8c6ed2a629ff357ff6cc73ed

Former-commit-id: 5d090b2c94f98f86cc55f1415927d287576169fa