Log warning if shade move request is not honored
As we're using "collectLatest", there are legit cases when the move
request is issued towards a display where the shade is at already (e.g.
the display id changed again before the shade completed the move, the
scope is cancelled, and we're calling moveShadeWindowTo again with the
original display)

We're still logging it as it might also signal some bug in the move.

Bug: 362719719
Test: ShadeDisplaysInteractor
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I0922a235d8a87ccd43170e38c2918ea3b10ed942

Former-commit-id: a0946e1029ccf321e898834c053b89967f848c16