Implement animation for the move-to-display feature
Previously, the move-to-display operation used the default animation.
With the port_window_size_animation feature flag enabled, the default
handler applies a resize animation. This is because move-to-display
transactions change the task size in pixels to handle density changes
and preserve the DP size. Although this is technically a pixel size
change, the user perceives that the DP size remains constant.

To avoid the resize animation from the default transition handler, this
CL implements a dedicated transition handler for move-to-display
operations.

Bug: 376355478
Test: DesktopTasksControllerTest
Test: DesktopModeMoveToDisplayTransitionHandlerTest
Flag: com.android.window.flags.enable_move_to_next_display_shortcut
Change-Id: I95100ccdb0b5b96b0a1e9943e75fcd75ea140ccc

Former-commit-id: d33ea49e196dd6b648a27ae600b6dc45832d1a49