WMCore-side infra for showing restart menu
This change introduces a new class AppCompatDisplayCompatModePolicy
, which encapsulates all our compat logic related to CD (WM part).
Currently, the class only owns the policy as to when the restart
menu should be shown. It targets all apps for now, and if they are
moved to a different display, the restart menu gets enabled, and
once they are restarted, it becomes invisible. This policy is
subject to change as we're finalizing the UX of the feature.

This policy is routed to WMShell via AppCompatTaskInfo and
accessible as isRestartMenuEnabledForDisplayMove() just like other
information on app-compat.

Flag: com.android.window.flags.enable_restart_menu_for_connected_displays
Bug: 400546294
Test: WmTests:DisplayCompatTests
Change-Id: I952c0dee9802a9288af170ff62d8818aec662aec

Former-commit-id: 41877c0ce26b1f810f4f60fff5c70b728751bc5b