[hotword] Fix async updateServiceIdentity
The callback used to updated the HDS identity is oneway, which means it
doesn't execute sync with respect to the other connection startup tasks,
and the service can be marked as ready before the identity is updated.

This can lead to some races with permission checks for HDS. Switch the
binder callback to non-oneway, and enqueue all the tasks on a dedicated
handler thread to avoid clogging the main thread.

Test: manual assistant enable/disable and functionality
Test: atest AlwaysOnHotwordDetectorTest
Bug: 384399959
Flag: EXEMPT bugfix
Change-Id: Ib527e28a1b0a0b2f7521a7ef505705db547eca9c

Former-commit-id: 2a383b6e9ebe9e416a6991efb464bb95e15b8835