Make bitmap ashmem filename unique
Bug: 369619160
Flag: NONE - bug fix

Expect bitmap ashmem filename to be unique so that we could identify the
bitmaps with the same ashmem allocations. However, in the case of a same
bitmap being sent multiple times, we could end up having the same ashmem
filenames from the receiving end.

To make the filename unique, an monotonically increasing counter is used
for ashmem allocations, and is included in the filename.

Change-Id: I08a170c4dbf465766d0303bc32f761e099dc2736

Former-commit-id: 86447050ec9beb3869843ee57ad2662681e94fae