Fix back callback not cleared by RIV
The onVisibilityAggregated function is not reliable for the RemoteInputView. In some cases it reports to be visible when it is not. That can cause back callbacks remain registered on the shade window back dispatcher. This CL changes the back callback registration/unregistration logic to be independent from onVisibilityAggregated.

Bug: 370628115
Flag: EXEMPT bugfix
Test: Manual, i.e. verified that reproduction steps no longer work to reproduce the bug.
Test: RemoteInputViewTest
Change-Id: I5ea0bb7ebe0a72bb2b0fea4012b007432841becd

Former-commit-id: 3db0392c347f9fb78cc07c43ae896287c9b9b595