Change ContentRevealTest to be smaller
Cuttlefish has a lower resolution display, which caused the
actual width to be less than what one gets on a test run on a local
device. This became a issue since not using the DeviceEmulationRule.
With this change, the regular device and CF goldens are compatible with
regular devices.

Also sneaking in a test change to ensure the
`floating_verticalReveal_gesture_magneticDetachAndReattach.json`
actually does detach (since teh floating versions's detach threshold is
higher)

Bug: 379612686
Test: Unit tests
Flag: TEST_ONLY
Change-Id: I8c0c8b4b6e27936fb11f00a3b824b31a19d89dd5

Former-commit-id: 4fd61a3ca10fb0c3740c89695bf912dda7616e58