Adding IME drawing status to ImeInsetsSourceProvider
During a rotation where the IME was and will be showing, it will redraw itself. However, we did not reset serverVisibility, which led to the control not being dispatched again, after the drawing is finished.
By storing the drawn status, we can reset server visibility when the IME is (temporarily) not drawn anymore.

Fix: 399240606
Flag: android.view.inputmethod.refactor_insets_controller
Test: atest com.android.server.wm.ImeInsetsSourceProviderTest#testOnPostLayout_resetServerVisibilityWhenImeIsNotDrawn
Test: manual:
      Open WhatsApp on tablet in landscape mode
      Open IME in chat
      Rotate; verify that IME is still showing after rotation
Change-Id: Ifc0d345633b1aede56cf9ca8b2517c051e7359c6

Former-commit-id: bfc916e4a1a256d75ebec3c5e2e42cb7bb6f775d