Fix collect extra intent keys on server
Previously, when keys are collected on the client side, intent is
marked as keys are collected as long as the bundle is unparceled.
But it is still possible that a Parcelable value is still parceled.
In such a case, if the Parcelable happens to be an intent, its key
is not collected, but the intent is marked as keys are collected.
As a result, the server side also won't try to collect it.
To fix it, remove the mark if there is any unparceled value.

Also, if the mExtras is null or definitely empty without having
to unparcel the bundle and check, mark the intent as keys are
collected.

Bug: 390317645
Test: manual
Flag: EXEMPT bugfix
Change-Id: I0e4bad042063c995364a4ba829cfc556b4698973

Former-commit-id: aa68eb995a1a5ed9830c4a14a49aa22c4901f451