[SB] Keep status bar hidden throughout the QQS -> QS transition.
`ShadeInteractor.isAnyFullyExpanded` is true once QQS is fully visible,
but as you do the QQS -> QS transition it becomes `false` during the
transition and `true` again at the end. (During the transition, neither
QQS nor QS is _fully_ visible)

The status bar should stay hidden throughout the QQS -> QS transition
because the shade is always covering it. So the status bar needs to
listen to a different `ShadeInteractor` flow that will stay `true`
throughout. In this CL, the status bar hides as long as at
least 20% of either shade *or* QS is visible (during the first 20% of
the nothing->QQS transition, the home status bar is still visible on
screen so we shouldn't hide it yet).

Bug: 394257529 (specifically comment24)
Bug: 364360986
Flag: com.android.systemui.status_bar_root_modernization
Test: Perform QQS -> QS transition, and QS -> QQS transition -> verify
status bar doesn't draw any frames
Test: atest HomeStatusBarViewModelImplTest

Change-Id: Ib1796c1aa2a32687f95c7dcea286b59d320a6026

Former-commit-id: 117dd373fceb2400d7abf1ddda08095a508c94db