Rename fadingBackground to animatedBackground with () -> Color param
This CL renames fadingBackground() to animatedBackground() and makes the
color parameter be a () -> Color, so that the color can also be
animated. This will be used by Expandable so that its color value is
read during the drawing phase.

The implementation is also now based on the Modifier Node API, and is
mostly a fork of the background() modifier implementation.

Bug: 402059261
Test: FooterActionsScreenshotTest
Flag: EXEMPT simple refactoring
Change-Id: Id4751ddb690b4c94d2cb669141cd27a2eb349b56

Former-commit-id: 22ed296e8ef8af495e43c5e69b4ab82713f3b344