Fix app icon of notifications posted by direct-boot-aware packages
To determine whether an app is "headless", we use getLaunchIntentForPackage; when this is called before the device is unlocked, it will only match direct-boot-aware launcher activities. However, it's fairly common for direct-boot-aware apps (such as DWB) to only have a direct-boot-unaware launcher activity, because they only intend to post notifications and not actually show UI until the user has unlocked the device.

Thus we add an overload to getLaunchIntentForPackage() that supports matching activities regardless of their direct-boot-awareness.

Bug: 401961480
Test: atest ApplicationPackageManagerTest
Flag: android.app.notifications_redesign_app_icons
Change-Id: Ie4cac3f8eeb92282655e9b6a870a13660e1e3c44

Former-commit-id: ef177d86282399abc1276b03c7f06cfe989067a0