Bring Internal and Public Histogram Buckets Inline
There was an issue mapping frame overruns that exceeded 1000ms, this was
due to the internal PendingJankStats containing one extra bucket for
histogram data vs what was available in the public
RelativeFrameTimeHistogram.

This change brings the bucket counts between the internal and public
objects inline and ensure the test that was failing is now passing.

Bug: 402159276
Test: atest CoreAppJankTests, CtsAppJankTestCases
Flag: EXEMPT bugfix
Change-Id: Ifaa1a18dd43792e0258c6b7818d41da8051d9738

Former-commit-id: 903d16e842373ad582da7a17649052546d17758a