[PiP on Desktop] Handle Desktop cleanups for all PiP transitions.
Previously, Desktop cleanup was only initiated when necessary (if task
going to PiP is the last task) when PiP is entered via the minimize
button on the app handle.

This CL updates the implementation so that PiP transitions detected
by PipTransition#handleRequest would also be checked, i.e. when PiP is
entered via back navigation.

For all cases, if Desktop session is active _and_ the task going to PiP
is the last task, the WCT for the PiP transition is modified to include
Desktop exit cleanups.

Bug: 359906472
Bug: 399976327
Test: atest DesktopPipTransitionControllerTest
DesktopTasksControllerTest
Flag: com.android.window.flags.enable_desktop_windowing_pip

Change-Id: I36fbc34a04ea418ce10683ad2cdae6681df1fcc9

Former-commit-id: fed05a65f5451c083564841eb60c48075fede261