[CD Testing] Ensure overlay displays are removed before and after tests
Note: Even if we want to restore the original overlay display added
before this test (and its topology), it will be complicated as re-adding
overlay display would lead to different displayId and topology could not
be restored easily. So it's preferred to just have a clean state before
and after the test.

Bug: 380001108
Test: presubmit
Flag: EXEMPT adding new tests
Change-Id: I6d1e09194990db6bd46af660cee80cc9c4b75621

Former-commit-id: 6da50226a2d7cbf243ee754d23c2127b0516e353