Add more conditions to excluding gesture
When QS is expanding in lockscreen or in split shade, there's no STL
transition, but BrightnessSlider is moving in the screen. This causes
the modifier to send a new rect (with global coordinates) on every
frame.

Instead, restrict this to only exclude gestures when QS is settled.

Test: manual, log calls to modifier lambda
Fixes: 401811481
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Change-Id: I3fc6754dc250524a3a40868a7ad557906a479d85

Former-commit-id: c277a107e27027a7167337d1b485905e69f69431