Disable zoom out effect when the shade is on an external display
The zoom out effect currently only supports the default display, but is triggered at every shade expansion (also when the shade is in an external display).

This looked unexpected. With this cl, the zoom out is only applied when the shade is in the default display.

Supporting the zoom out on the external display requires some refactor of AppZoomOutController (tracked in b/403573903)

Bug: 362719719
Bug: 401606287
Bug: 403573903
Test: NotificationShadeDepthControllerTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I1d82dbc074ca2c23af03cb9414b13d685a9259bc

Former-commit-id: 045c900873524c59472123bdee80e5f148fc75f8