Playing device entry haptics from the DevieEntryIcon.
Whenever entering the device via the DeviceEntryIcon, we also play the
device entry haptics. This keeps the CUJ consistent with all the other
forms of haptic feedback when entering the device. We also keep the
previous behavior consistent when pressing the icon opens the bouncer
instead of entering the device. In such a case, the long-press MSDL
feedback token plays.

Test: DeviceEntryHapticsInteractorTest
Flag: com.android.systemui.msdl_feedback
Bug: 403255619
Change-Id: I1d6f985c42a17b0b53e3a26ae673c4ba9f3919a4

Former-commit-id: 156afbb66a9879b3d3d40ecad747778f06001743