Fixed an issue where offsets of animations could get stuck
This would lead to overlapping children.
When group notificaitons are added to their parents, all
running animations are cancelled. Since we're now animating
offsets and never again set such offsets unless we animate,
We could get stuck with such an offset, continuously added
to the regular translation.
We now always reset the offset to 0 when ending it as that
seems to be the safest thing to do

Fixes: 402273107
Flag: com.android.systemui.physical_notification_movement
Test: atest SystemUiTests
Change-Id: Ie2494395549a9329252331a4fa167381550d90a7

Former-commit-id: c21f5c58291b8d29fba9f292ed10fad4898463e6