Do not remove decoration surface on DragResizeInputListener#close
I37ba7d97d51079ba07148a425826a0cd9b59c3fd made a change to create a copy
of the decoration container surface to allow DRIL to manage its release
in the correct thread. However, it started using SC.T#remove() to
release the surface, which is fine in most cases because closing the
DRIL usually happens when the decoration is also closing. In
desktop-immersive cases though, where we make a desktop task use
fullscreen bounds, we want to dispose of the DRIL because it is no
longer drag-resizable, but we want to keep the decoration surface
because we still show an App Header.

With this change, the copied decoration surface is only release in
close(), and removal continues to be managed by WindowDecoration.

Flag: com.android.window.flags.enable_drag_resize_set_up_in_bg_thread
Fix: 404581880
Test: put immersive-capable app into desktop-immersive, drag down from
status bar - verify App Header becomes visible

Change-Id: I029c56189fc8656789cf0ece2437d8cc16475208

Former-commit-id: 33f0ce08a5408ef1bafd0067ffc334ab28d2d2d4