[Bugfix]Use local width and height to get cutout in the target direction
In the vertical-to-horizontal screen orientation switch scenario, when SystemUI obtains the cutout through method applyHeight, using logicWidth and logicHeight may result in incorrect calculation of the rect for devices configured with side cutouts, leading to abnormal insets updates due to the Display updating the horizontal screen information in advance.
By aligning the direction transformation calculation with the current local width and height, the issue of abnormal calculation of side cutouts in rotation scenarios can be fundamentally resolved.

Bug: 404727504
Test:See the bug's reproduce steps.
Change-Id: I499c73a09610bdd941afe3367112dfc1943f747c

Former-commit-id: 7d432029f2b7315020aea8c3233b1277e89a3d10