[res] Remove global overlays from registered lib paths
When calling Resources.registerResourcePaths(), the passed
application info object contains all asset paths for the lib,
including global system overlays, theming FRROs and so on.

This change filters out those that are common with the current
app's assets - they are already a part of all resource objects
and don't need to be registered and added separately.

This fixes an issue of retaining global overlay paths that got
updated after the register...() function call

Bug: 383713434
Test: unit tests + resources CTS
Flag: EXEMPT bugfix
Change-Id: Ife713e364c3e4b68a9e8c480f3773df40a0cc39c

Former-commit-id: 0a856ff1a9bb281f5ee62b2fe1f42966f233b388