Permissions: Sync BodySensors/ReadHr permissions on app install.
If either permission is not granted, revoke the other one. This forces the user to re-grant the permission with the new scope. Also validates that the body-sensor background permission is only granted when the foreground permission is granted as well.

This helps to ensure that as apps are installed on the latest platform that they can continue using both new health and legacy body sensors permissions together in a backwards compatible way.

Bug: 401614607
Test: Manually built and verified some syncing scenarios with test app.
Test: atest PermissionServiceMockingTests
Flag: EXEMPT bugfix
Change-Id: I8be692527e912c5e90336f8fb88cabd07efe601c

Former-commit-id: 051b999dcfd8439c0e215a298f18ee413e966a46